ASU-546: moved application event under application module
